+++ src/platform/assets/index.ts
@@ -1,0 +1,3 @@
+// Public API exports for asset platform
+export * from './schemas/assetSchema'
+export * from './services/assetService'
