+++ src/platform/assets/schemas/assetSchema.ts
