+++ src/renderer/extensions/vueNodes/widgets/composables/useComboWidget.ts
@@ -7,6 +7,7 @@ import type {
   IBaseWidget,
   IComboWidget
 } from '@/lib/litegraph/src/types/widgets'
+import { assetService } from '@/platform/assets'
 import { useSettingStore } from '@/platform/settings/settingStore'
 import { transformInputSpecV2ToV1 } from '@/schemas/nodeDef/migration'
 import {
@@ -23,7 +24,6 @@ import {
   type ComfyWidgetConstructorV2,
   addValueControlWidgets
 } from '@/scripts/widgets'
-import { assetService } from '@/services/assetService'
 
 import { useRemoteWidget } from './useRemoteWidget'
