+++ src/scripts/api.ts
@@ -2,6 +2,7 @@ import axios from 'axios'
 import { get } from 'es-toolkit/compat'
 
 import defaultClientFeatureFlags from '@/config/clientFeatureFlags.json' with { type: 'json' }
+import type { ModelFile, ModelFolderInfo } from '@/platform/assets'
 import { useToastStore } from '@/platform/updates/common/toastStore'
 import { WorkflowTemplates } from '@/platform/workflow/templates/types/template'
 import type {
@@ -37,7 +38,6 @@ import type {
   User,
   UserDataFullInfo
 } from '@/schemas/apiSchema'
-import type { ModelFile, ModelFolderInfo } from '@/schemas/assetSchema'
 import type { ComfyNodeDef } from '@/schemas/nodeDefSchema'
 import type { NodeExecutionId } from '@/types/nodeIdentification'
